refactor(#155 PR 4/10): extract core/actions.js and core/events.js from game.js
Pure code movement, zero behavior change. game.js: 4076 -> 3374 lines.

- src/core/actions.js (262 lines): the request lifecycle — spawnRequest,
  routeRequestToEntry, finish/fail/throttle/removeRequest, updateScore(+UI),
  flashMoney, getUpkeepMultiplier, calculateFailChanceBasedOnLoad and their
  private helpers.
- src/core/events.js (490 lines): the random-events system — intervention
  warnings, malicious spikes, traffic shifts, event timers.
- Service.js / Request.js / campaign.js now import these from core/ instead
  of game.js, eliminating most of the game.js import cycles; game.js exports
  shrink to formatTime + scene groups + two campaign UI hooks.
- tests/i18n-usage scans src/core too, keeping moved keys under audit.

Verified: lint 0 (no-undef on), 87/87 tests, full module graph links
(24 modules), and in-browser — campaign level 2 prebuilds, 50/50 requests
processed through the moved spawnRequest with campaign counters advancing,
events fire from the new module, campaign tick clean, save-file upload
round-trip works, zero console errors.

Note on PR 3 (locales->JSON): skipped — the validator goal is already
CI-enforced by the vitest locale suite, and JSON modules would require
import attributes that Firefox ESR still lacks, breaking raw hosting for
real users. Documented in #155.
