+++ game.js
@@ -10,6 +10,23 @@ import "./src/tutorial.js";
 import "./src/campaign/campaign.js";
 import { CAMPAIGN_LEVELS } from "./src/campaign/levels.js";
 import { renderArchitectureSVG } from "./src/campaign/diagram.js";
+import {
+    flashMoney,
+    getUpkeepMultiplier,
+    removeRequest,
+    routeRequestToEntry,
+    spawnRequest,
+    updateScoreUI,
+} from "./src/core/actions.js";
+import {
+    addInterventionWarning,
+    endRandomEvent,
+    triggerRandomEvent,
+    updateActiveEventTimer,
+    updateMaliciousSpike,
+    updateRandomEvents,
+    updateTrafficShift,
+} from "./src/core/events.js";
 
 STATE.sound = new SoundService();
 
@@ -66,316 +83,6 @@ function calculateTargetRPS(gameTimeSeconds) {
     return targetRPS;
 }
 
-function getUpkeepMultiplier() {
-    if (STATE.gameMode !== "survival") return 1.0;
-    if (!CONFIG.survival.upkeepScaling.enabled) return 1.0;
-
-    const gameTime =
-        STATE.elapsedGameTime ?? (performance.now() - STATE.gameStartTime) / 1000;
-    const progress = Math.min(
-        gameTime / CONFIG.survival.upkeepScaling.scaleTime,
-        1.0
-    );
-
-    const base = CONFIG.survival.upkeepScaling.baseMultiplier;
-    const max = CONFIG.survival.upkeepScaling.maxMultiplier;
-
-    let multiplier = base + (max - base) * progress;
-
-    if (STATE.intervention?.costMultiplier) {
-        multiplier *= STATE.intervention.costMultiplier;
-    }
-
-    return multiplier;
-}
-
-function updateMaliciousSpike(dt) {
-    if (STATE.gameMode === "campaign") {
-        if (!STATE.campaign?.level?.enableSurvivalShifts) return;
-    } else if (STATE.gameMode !== "survival") {
-        return;
-    }
-    if (!CONFIG.survival.maliciousSpike.enabled) return;
-
-    STATE.maliciousSpikeTimer += dt;
-
-    const interval = CONFIG.survival.maliciousSpike.interval;
-    const duration = CONFIG.survival.maliciousSpike.duration;
-    const warning = CONFIG.survival.maliciousSpike.warningTime;
-
-    const cycleTime = STATE.maliciousSpikeTimer % interval;
-
-    if (
-        cycleTime >= interval - warning &&
-        cycleTime < interval - warning + dt &&
-        !STATE.maliciousSpikeActive
-    ) {
-        showMaliciousWarning();
-    }
-
-    if (cycleTime < dt && STATE.maliciousSpikeTimer > warning) {
-        startMaliciousSpike();
-    }
-
-    if (
-        STATE.maliciousSpikeActive &&
-        cycleTime >= duration &&
-        cycleTime < duration + dt
-    ) {
-        endMaliciousSpike();
-    }
-}
-
-function showMaliciousWarning() {
-    const existing = document.getElementById("malicious-warning");
-    if (existing) existing.remove();
-
-    const warning = document.createElement("div");
-    warning.id = "malicious-warning";
-    warning.className =
-        "fixed top-1/3 left-1/2 transform -translate-x-1/2 text-center z-50 pointer-events-none";
-    warning.innerHTML = `
-        <div class="text-red-500 text-2xl font-bold animate-pulse">${i18n.t('ddos_incoming')}</div>
-        <div class="text-red-300 text-sm">${i18n.t('attack_spike')}</div>
-    `;
-    document.body.appendChild(warning);
-
-    STATE.sound.playTone(400, "sawtooth", 0.3);
-    STATE.sound.playTone(300, "sawtooth", 0.3, 0.15);
-
-    setTimeout(() => warning.remove(), 4000);
-}
-
-function startMaliciousSpike() {
-    const existing = document.getElementById("malicious-spike-indicator");
-    if (existing) existing.remove();
-
-    if (STATE.intervention && STATE.intervention.trafficShiftActive) return;
-
-    STATE.maliciousSpikeActive = true;
-
-    STATE.normalTrafficDist = { ...STATE.trafficDistribution };
-
-    const maliciousPct = CONFIG.survival.maliciousSpike.maliciousPercent;
-    const remaining = 1 - maliciousPct;
-
-    // Guard against a distribution that's already 100% malicious — otherwise
-    // every non-malicious share divides by zero and becomes NaN/Infinity,
-    // corrupting the mix for the spike's duration. Not reachable with shipped
-    // configs today, but cheap insurance for future levels/shifts.
-    const otherTotal = 1 - STATE.normalTrafficDist.MALICIOUS;
-    if (otherTotal <= 0) {
-        STATE.trafficDistribution = { ...STATE.normalTrafficDist };
-    } else {
-        STATE.trafficDistribution = {
-            STATIC: (STATE.normalTrafficDist.STATIC / otherTotal) * remaining,
-            READ: (STATE.normalTrafficDist.READ / otherTotal) * remaining,
-            WRITE: (STATE.normalTrafficDist.WRITE / otherTotal) * remaining,
-            UPLOAD: (STATE.normalTrafficDist.UPLOAD / otherTotal) * remaining,
-            SEARCH: (STATE.normalTrafficDist.SEARCH / otherTotal) * remaining,
-            MALICIOUS: maliciousPct,
-        };
-    }
-
-    const indicator = document.createElement("div");
-    indicator.id = "malicious-spike-indicator";
-    indicator.className =
-        "fixed top-4 left-1/2 transform -translate-x-1/2 z-40 pointer-events-none";
-    indicator.innerHTML = `
-        <div class="bg-red-900/80 border-2 border-red-500 rounded-lg px-4 py-2 animate-pulse">
-            <span class="text-red-400 font-bold">${i18n.t('ddos_active')}</span>
-        </div>
-    `;
-    document.body.appendChild(indicator);
-
-    const maliciousEl = document.getElementById("mix-malicious");
-    if (maliciousEl)
-        maliciousEl.className = "text-red-500 font-bold animate-pulse";
-}
-
-function endMaliciousSpike() {
-    STATE.maliciousSpikeActive = false;
-
-    // Restore normal distribution
-    if (STATE.normalTrafficDist) {
-        STATE.trafficDistribution = { ...STATE.normalTrafficDist };
-        STATE.normalTrafficDist = null;
-    }
-
-    // Remove indicator
-    const indicator = document.getElementById("malicious-spike-indicator");
-    if (indicator) indicator.remove();
-
-    // Reset mix display styling
-    const maliciousEl = document.getElementById("mix-malicious");
-    if (maliciousEl) maliciousEl.className = "text-red-400";
-
-    STATE.sound.playSuccess();
-}
-
-// ==================== INTERVENTION MECHANICS ====================
-
-function addInterventionWarning(message, type = "warning", duration = 4000) {
-    const warningsContainer = document.getElementById("intervention-warnings");
-    if (!warningsContainer) return;
-
-    const warning = document.createElement("div");
-    const typeStyles = {
-        warning: "warning-warning",
-        danger: "warning-danger",
-        info: "warning-info",
-    };
-
-    warning.className = `intervention-warning ${typeStyles[type] || typeStyles.warning
-        } border-2 rounded-lg px-6 py-3 mb-2 shadow-lg`;
-    warning.innerHTML = `
-        <div class="flex items-center gap-3">
-            <span class="text-2xl">${type === "danger" ? "⚠️" : type === "info" ? "✅" : "📢"
-        }</span>
-            <span class="font-bold text-lg">${message}</span>
-        </div>
-    `;
-    warningsContainer.appendChild(warning);
-
-    // Play warning sound
-    if (type === "danger") {
-        STATE.sound?.playTone(200, "sawtooth", 0.4);
-        STATE.sound?.playTone(150, "sawtooth", 0.4, 0.1);
-    } else if (type === "warning") {
-        STATE.sound?.playTone(400, "sine", 0.2);
-    }
-
-    // Add to state for tracking
-    if (STATE.intervention) {
-        STATE.intervention.warnings.push({ message, type, time: Date.now() });
-    }
-
-    // Animate out before removing
-    setTimeout(() => {
-        warning.style.transition = "all 0.3s ease-out";
-        warning.style.opacity = "0";
-        warning.style.transform = "translateY(-20px)";
-        setTimeout(() => warning.remove(), 300);
-    }, duration - 300);
-}
-
-function updateTrafficShift(dt) {
-    if (STATE.gameMode === "campaign") {
-        if (!STATE.campaign?.level?.enableSurvivalShifts) return;
-    } else if (STATE.gameMode !== "survival") {
-        return;
-    }
-    if (!CONFIG.survival.trafficShift?.enabled) return;
-    if (!STATE.intervention) return;
-
-    STATE.intervention.trafficShiftTimer += dt;
-
-    const config = CONFIG.survival.trafficShift;
-    const interval = config.interval;
-    const duration = config.duration;
-
-    // Check if shift should start
-    if (
-        !STATE.intervention.trafficShiftActive &&
-        STATE.intervention.trafficShiftTimer >= interval
-    ) {
-        startTrafficShift();
-    }
-
-    // Check if shift should end. The timer is reset to 0 when a shift actually
-    // activates (see startTrafficShift), so here it measures time-since-active.
-    // Previously this used the absolute `interval + duration` threshold, which
-    // meant a shift that was delayed (blocked while a malicious spike was active)
-    // could end on its very first active frame — running for ~0 seconds.
-    if (
-        STATE.intervention.trafficShiftActive &&
-        STATE.intervention.trafficShiftTimer >= duration
-    ) {
-        endTrafficShift();
-        STATE.intervention.trafficShiftTimer = 0; // Reset for next cycle
-    }
-}
-
-function startTrafficShift() {
-    if (!STATE.intervention || STATE.maliciousSpikeActive) return;
-
-    const config = CONFIG.survival.trafficShift;
-    const shifts = config.shifts;
-
-    // Pick a random shift
-    const shift = shifts[Math.floor(Math.random() * shifts.length)];
-    STATE.intervention.currentShift = shift;
-    STATE.intervention.trafficShiftActive = true;
-    // Reset so the end check measures duration from actual activation, not from
-    // when the timer first crossed `interval` (it may have kept growing while a
-    // malicious spike blocked the start).
-    STATE.intervention.trafficShiftTimer = 0;
-
-    // Store original distribution
-    STATE.intervention.originalTrafficDist = { ...STATE.trafficDistribution };
-
-    if (shift.distribution) {
-        STATE.trafficDistribution = { ...shift.distribution };
-    }
-
-    // Shift configs carry only { name, distribution } — there is no `type`
-    // field, and most shift names have no `shift_<name>` i18n key. Referencing
-    // shift.type threw on every shift start (crashing the frame and suppressing
-    // this warning). Use the shift's display name directly.
-    addInterventionWarning(
-        i18n.t('traffic_surging', { name: shift.name }),
-        "warning",
-        5000
-    );
-    STATE.sound?.playTone(500, "sine", 0.2);
-}
-
-function endTrafficShift() {
-    if (!STATE.intervention) return;
-
-    STATE.intervention.trafficShiftActive = false;
-
-    // Restore original distribution
-    if (STATE.intervention.originalTrafficDist) {
-        STATE.trafficDistribution = { ...STATE.intervention.originalTrafficDist };
-        STATE.intervention.originalTrafficDist = null;
-    }
-
-    STATE.intervention.currentShift = null;
-}
-
-function updateRandomEvents(dt) {
-    if (STATE.gameMode === "campaign") {
-        if (!STATE.campaign?.level?.enableSurvivalShifts) return;
-    } else if (STATE.gameMode !== "survival") {
-        return;
-    }
-    if (!CONFIG.survival.randomEvents?.enabled) return;
-    if (!STATE.intervention) return;
-
-    STATE.intervention.randomEventTimer += dt;
-
-    const config = CONFIG.survival.randomEvents;
-
-    // Check if event should trigger
-    if (STATE.intervention.randomEventTimer >= config.checkInterval) {
-        STATE.intervention.randomEventTimer = 0;
-
-        // 30% chance to trigger an event
-        if (Math.random() < 0.3) {
-            triggerRandomEvent();
-        }
-    }
-
-    // Check if active event should end
-    if (
-        STATE.intervention.activeEvent &&
-        Date.now() >= STATE.intervention.eventEndTime
-    ) {
-        endRandomEvent();
-    }
-}
-
 window.handleGameState = (timeScale) => {
     if (timeScale === 0) { // pause state
         STATE.intervention.pausedEvent = STATE.intervention.activeEvent;
@@ -395,189 +102,6 @@ window.handleGameState = (timeScale) => {
     }
 
     window.setTimeScale(timeScale);
-}
-
-function triggerRandomEvent(
-    eventType = null,
-    duration = null,
-    outageServiceId = null
-) {
-    if (!STATE.intervention || STATE.intervention.activeEvent) return;
-
-    const config = CONFIG.survival.randomEvents;
-    if (!eventType)
-        eventType = config.types[Math.floor(Math.random() * config.types.length)];
-    if (!duration) duration = 30000; // 30 seconds
-
-    STATE.intervention.activeEvent = eventType;
-    STATE.intervention.eventEndTime = Date.now() + duration;
-    STATE.intervention.eventDuration = duration;
-
-    switch (eventType) {
-        case "COST_SPIKE":
-            addInterventionWarning(
-                i18n.t('cost_spike_warning'),
-                "danger",
-                8000
-            );
-            STATE.intervention.costMultiplier = 2.0;
-            break;
-
-        case "CAPACITY_DROP":
-            addInterventionWarning(
-                i18n.t('capacity_drop_warning'),
-                "danger",
-                8000
-            );
-            STATE.services.forEach((s) => {
-                s.tempCapacityReduction = 0.5; // 50% capacity
-            });
-            break;
-
-        case "TRAFFIC_BURST":
-            addInterventionWarning(
-                i18n.t('traffic_burst_warning'),
-                "warning",
-                8000
-            );
-            STATE.intervention.trafficBurstMultiplier = 3.0;
-            break;
-
-        case "SERVICE_OUTAGE": {
-            // Reuse the previously-chosen service when resuming a paused outage
-            // (STATE.intervention.outageServiceId set), otherwise pick a fresh
-            // random one. Without this, pause→resume re-rolled the target and
-            // the outage could "teleport" to a different service.
-            let target = outageServiceId
-                ? STATE.services.find((s) => s.id === outageServiceId)
-                : null;
-            if (!target) {
-                const services = STATE.services.filter((s) => s.type !== "waf");
-                target = services.length > 0
-                    ? services[Math.floor(Math.random() * services.length)]
-                    : null;
-            }
-            if (target) {
-                STATE.intervention.outageServiceId = target.id;
-                target.isDisabled = true;
-                target.mesh.material.opacity = 0.3;
-                target.mesh.material.transparent = true;
-                addInterventionWarning(
-                    i18n.t('service_outage_warning', { type: i18n.t(target.type) }),
-                    "danger",
-                    8000
-                );
-            }
-            break;
-        }
-    }
-
-    // Show active event bar
-    showActiveEventBar(eventType);
-
-    STATE.sound?.playTone(300, "sawtooth", 0.3);
-}
-
-function endRandomEvent() {
-    if (!STATE.intervention || !STATE.intervention.activeEvent) return;
-
-    const eventType = STATE.intervention.activeEvent;
-
-    switch (eventType) {
-        case "COST_SPIKE":
-            STATE.intervention.costMultiplier = 1.0;
-            break;
-
-        case "CAPACITY_DROP":
-            STATE.services.forEach((s) => {
-                s.tempCapacityReduction = 1.0;
-            });
-            break;
-
-        case "TRAFFIC_BURST":
-            STATE.intervention.trafficBurstMultiplier = 1.0;
-            break;
-
-        case "SERVICE_OUTAGE":
-            STATE.services.forEach((s) => {
-                if (s.isDisabled) {
-                    s.isDisabled = false;
-                    s.mesh.material.opacity = 1.0;
-                    s.mesh.material.transparent = false;
-                }
-            });
-            // Clear the remembered target. On a pause, handleGameState has
-            // already captured it into pausedOutageServiceId before calling this,
-            // so clearing here only affects a genuine end — the next fresh outage
-            // will pick a new service.
-            STATE.intervention.outageServiceId = null;
-            break;
-    }
-
-    // Hide active event bar
-    hideActiveEventBar();
-
-    STATE.intervention.activeEvent = null;
-    addInterventionWarning(i18n.t('event_ended'), "info", 2000);
-    STATE.sound?.playSuccess();
-}
-
-function showActiveEventBar(eventType) {
-    const bar = document.getElementById("active-event-bar");
-    const icon = document.getElementById("active-event-icon");
-    const text = document.getElementById("active-event-text");
-
-    if (!bar) return;
-
-    const eventConfig = {
-        COST_SPIKE: { icon: "💰", text: i18n.t('cost_spike_active'), color: "bg-red-600" },
-        CAPACITY_DROP: {
-            icon: "⚡",
-            text: i18n.t('capacity_reduced'),
-            color: "bg-orange-600",
-        },
-        TRAFFIC_BURST: {
-            icon: "🚀",
-            text: i18n.t('traffic_burst'),
-            color: "bg-yellow-600",
-        },
-        SERVICE_OUTAGE: {
-            icon: "🔧",
-            text: i18n.t('service_outage_active'),
-            color: "bg-purple-600",
-        },
-    };
-
-    const config = eventConfig[eventType] || eventConfig["COST_SPIKE"];
-
-    bar.className = `fixed top-0 left-0 right-0 h-8 z-40 ${config.color}`;
-    icon.textContent = config.icon;
-    text.textContent = config.text;
-    bar.classList.remove("hidden");
-}
-
-function hideActiveEventBar() {
-    const bar = document.getElementById("active-event-bar");
-    if (bar) bar.classList.add("hidden");
-}
-
-function updateActiveEventTimer() {
-    if (!STATE.intervention?.activeEvent) return;
-
-    const timerEl = document.getElementById("active-event-timer");
-    const progressEl = document.getElementById("active-event-progress");
-
-    const remaining = Math.max(0, STATE.intervention.eventEndTime - Date.now());
-    const remainingSec = Math.ceil(remaining / 1000);
-
-    if (timerEl) {
-        timerEl.textContent = formatTime(remainingSec);
-    }
-
-    if (progressEl && STATE.intervention.eventDuration) {
-        const progress = (remaining / STATE.intervention.eventDuration) * 100;
-        progressEl.style.width = `${Math.max(0, progress)}%`;
-    }
 }
 
 function updateServiceHealthIndicators() {
@@ -1462,212 +986,6 @@ function snapToGrid(vec) {
     );
 }
 
-function getTrafficType() {
-    const dist = STATE.trafficDistribution;
-    const types = Object.keys(dist);
-    const total = types.reduce((sum, type) => sum + (dist[type] || 0), 0);
-    // All types at 0% means "no traffic", not "default to STATIC" (#174).
-    if (total === 0) return null;
-
-    const r = Math.random() * total;
-    let cumulative = 0;
-
-    for (const type of types) {
-        cumulative += dist[type] || 0;
-        if (r < cumulative) {
-            return TRAFFIC_TYPES[type] || type;
-        }
-    }
-
-    return TRAFFIC_TYPES.STATIC;
-}
-
-// Round-robin counters for entry-point load splitting across
-// multiple services of the same type (e.g. two WAFs on the Internet).
-// Keyed by service type ("waf", "cdn", "apigw", "any").
-const entryRRIndex = {};
-
-function pickEntryNode(entryNodes, type) {
-    // Filter for live (non-disabled) nodes of the requested type.
-    // Type "any" means "any live entry node" (last-resort path).
-    const candidates = entryNodes.filter((s) => {
-        if (!s || s.isDisabled) return false;
-        return type === "any" ? true : s.type === type;
-    });
-    if (candidates.length === 0) return null;
-    if (candidates.length === 1) return candidates[0];
-
-    // Round robin: each subsequent call rotates to the next candidate,
-    // splitting load evenly across identical entry points.
-    const idx = (entryRRIndex[type] || 0) % candidates.length;
-    entryRRIndex[type] = idx + 1;
-    return candidates[idx];
-}
-
-function spawnRequest() {
-    const type = getTrafficType();
-    // No traffic mix configured (all sliders at 0%) — nothing to spawn (#174).
-    if (type === null) return;
-    const req = new Request(type);
-    STATE.requests.push(req);
-    routeRequestToEntry(req, type);
-}
-
-// Shared entry routing for spawned traffic (regular spawns AND sandbox bursts).
-// Round-robin aware so multiple firewalls / CDNs / gateways share the load.
-function routeRequestToEntry(req, type) {
-    const conns = STATE.internetNode.connections;
-    if (conns.length === 0) {
-        failRequest(req);
-        return;
-    }
-    const entryNodes = conns.map((id) =>
-        STATE.services.find((s) => s.id === id)
-    );
-
-    let target;
-
-    // 1. Prefer CDN for STATIC traffic
-    if (type === "STATIC") {
-        target = pickEntryNode(entryNodes, "cdn");
-    }
-
-    // 2. Fallback to WAF (Security Best Practice)
-    if (!target) {
-        target = pickEntryNode(entryNodes, "waf");
-    }
-
-    // 3. Fallback to API Gateway (Rate Limiting)
-    if (!target) {
-        target = pickEntryNode(entryNodes, "apigw");
-    }
-
-    // 4. Last Resort: any live entry point (also round-robin)
-    if (!target) {
-        target = pickEntryNode(entryNodes, "any");
-    }
-
-    if (target) req.flyTo(target);
-    else failRequest(req);
-}
-
-function updateScore(req, outcome) {
-    const points = CONFIG.survival.SCORE_POINTS;
-    const typeConfig = req.typeConfig || CONFIG.trafficTypes[req.type];
-
-    if (outcome === "MALICIOUS_BLOCKED") {
-        STATE.score.maliciousBlocked += points.MALICIOUS_BLOCKED_SCORE;
-        STATE.score.total += points.MALICIOUS_BLOCKED_SCORE;
-
-        // Mitigation cost for blocking attacks
-        const mitigationCost = CONFIG.survival.SCORE_POINTS.MALICIOUS_MITIGATION_COST || 1.0;
-        STATE.money -= mitigationCost;
-        if (STATE.finances) {
-            STATE.finances.expenses.mitigation = (STATE.finances.expenses.mitigation || 0) + mitigationCost;
-        }
-        STATE.sound.playFraudBlocked();
-    } else if (
-        req.type === TRAFFIC_TYPES.MALICIOUS &&
-        outcome === "MALICIOUS_PASSED"
-    ) {
-        STATE.reputation += points.MALICIOUS_PASSED_REPUTATION;
-        STATE.failures.MALICIOUS++;
-
-        // Breach penalty
-        const breachPenalty = CONFIG.survival.SCORE_POINTS.MALICIOUS_BREACH_PENALTY || 50.0;
-        STATE.money -= breachPenalty;
-        if (STATE.finances) {
-            STATE.finances.expenses.breach = (STATE.finances.expenses.breach || 0) + breachPenalty;
-        }
-
-        console.warn(
-            `MALICIOUS PASSED: ${points.MALICIOUS_PASSED_REPUTATION} Rep. (Critical Failure)`
-        );
-    } else if (outcome === "COMPLETED") {
-        let reward = typeConfig.reward;
-        const score = typeConfig.score;
-
-        if (req.cached) {
-            reward *= 1 + points.CACHE_HIT_BONUS;
-        }
-
-        if (typeConfig.destination === "s3" || typeConfig.destination === "cdn") {
-            STATE.score.storage += score;
-        } else if (typeConfig.destination === "db") {
-            STATE.score.database += score;
-        }
-
-        STATE.score.total += score;
-        STATE.money += reward;
-        if (STATE.finances) {
-            STATE.finances.income.requests += reward;
-            STATE.finances.income.total += reward;
-            // Track by request type
-            const reqType = req.type || "STATIC";
-            STATE.finances.income.byType[reqType] =
-                (STATE.finances.income.byType[reqType] || 0) + reward;
-            STATE.finances.income.countByType[reqType] =
-                (STATE.finances.income.countByType[reqType] || 0) + 1;
-        }
-        STATE.reputation += points.SUCCESS_REPUTATION || 0.5; // Gain reputation on success
-    } else if (outcome === "THROTTLED") {
-        // Soft fail from API Gateway rate limiting — much less reputation loss
-        STATE.reputation += points.THROTTLED_REPUTATION || -0.2;
-    } else if (outcome === "FAILED") {
-        STATE.reputation += points.FAIL_REPUTATION;
-        STATE.score.total -= (typeConfig.score || 5) / 2;
-        if (STATE.failures[req.type] !== undefined) {
-            STATE.failures[req.type]++;
-        }
-    }
-
-    updateScoreUI();
-}
-
-function finishRequest(req, viaServiceType) {
-    STATE.requestsProcessed++;
-    updateScore(req, "COMPLETED");
-    if (window.campaign?.active) {
-        window.campaign.onRequestCompleted(req, viaServiceType);
-    }
-    removeRequest(req);
-}
-
-function failRequest(req) {
-    const failType =
-        req.type === TRAFFIC_TYPES.MALICIOUS ? "MALICIOUS_PASSED" : "FAILED";
-    updateScore(req, failType);
-    STATE.sound.playFail();
-    req.mesh.material.color.setHex(CONFIG.colors.requestFail);
-    setTimeout(() => removeRequest(req), 500);
-}
-
-function throttleRequest(req) {
-    updateScore(req, "THROTTLED");
-    STATE.sound.playFail();
-    req.mesh.material.color.setHex(CONFIG.colors.apigw); // Pink flash for throttled
-    setTimeout(() => removeRequest(req), 500);
-}
-
-function removeRequest(req) {
-    req.destroy();
-    STATE.requests = STATE.requests.filter((r) => r !== req);
-}
-
-function updateScoreUI() {
-    document.getElementById("total-score-display").innerText = STATE.score.total;
-    document.getElementById("score-storage").innerText = STATE.score.storage;
-    document.getElementById("score-database").innerText = STATE.score.database;
-    document.getElementById("score-malicious").innerText =
-        STATE.score.maliciousBlocked;
-}
-
-function flashMoney() {
-    const el = document.getElementById("money-display");
-    el.classList.add("text-red-500");
-    setTimeout(() => el.classList.remove("text-red-500"), 300);
-}
-
 function showMainMenu() {
     // Ensure sound is initialized if possible (browsers might block until interaction)
     if (!STATE.sound.ctx) STATE.sound.init();
@@ -2298,16 +1616,6 @@ function deleteObject(id) {
     STATE.money += Math.floor(svc.config.cost / 2);
     STATE.sound.playDelete();
     updateRepairCostTable();
-}
-
-/**
- * Calculates the percentage if failure based on the load of the node.
- * @param {number} load fractions of 1 (0 to 1) of how loaded the node is
- * @returns {number} chance of failure (0 to 1)
- */
-function calculateFailChanceBasedOnLoad(load) {
-    if (load <= 0.5) return 0;
-    return 2 * (load - 0.5);
 }
 
 window.setTool = (t) => {
@@ -4054,22 +3362,13 @@ window.toggleAutoRepair = toggleAutoRepair;
 // binding. Keep STATE reachable from there.
 window.STATE = STATE;
 
-// Runtime cross-module surface: Request.js, Service.js and campaign.js import
-// these (cyclically — safe, they are hoisted declarations / top-level consts
-// only dereferenced after evaluation).
+// Runtime cross-module surface: Request.js, Service.js, campaign.js and
+// core/events.js import these (cyclically — safe, they are hoisted
+// declarations / top-level consts only dereferenced after evaluation).
 export {
-    addInterventionWarning,
-    calculateFailChanceBasedOnLoad,
-    failRequest,
-    finishRequest,
-    flashMoney,
-    getUpkeepMultiplier,
-    removeRequest,
+    formatTime,
     renderCampaignObjectives,
     requestGroup,
     serviceGroup,
     showCampaignDebrief,
-    spawnRequest,
-    throttleRequest,
-    updateScore,
 };
