+++ src/entities/Request.js
@@ -1,9 +1,10 @@
 import { CONFIG } from "../config.js";
 import { STATE } from "../state.js";
-// Cyclic import (game.js ⇄ Request.js) is safe: these are hoisted function
-// declarations / top-level consts in game.js, only dereferenced at runtime —
-// long after both modules have finished evaluating.
-import { failRequest, requestGroup } from "../../game.js";
+// Cyclic imports (game.js / core/actions.js ⇄ Request.js) are safe: these are
+// hoisted function declarations / top-level consts, only dereferenced at
+// runtime — long after all modules have finished evaluating.
+import { failRequest } from "../core/actions.js";
+import { requestGroup } from "../../game.js";
 
 export class Request {
     constructor(type) {
